### 2021-05-24 #116 @whedon
[PRE REVIEW]: Developing and deploying an integrated workshop curriculum teaching computational skills for reproducible research
**Submitting author:** @pschloss (<a href="http://orcid.org/0000-0002-6935-4275">Patrick D Schloss</a>)
**Repository:** <a href="https://github.com/UMCarpentries/intro-curriculum-r/tree/gh-pages/paper" target ="_blank">https://github.com/UMCarpentries/intro-curriculum-r/tree/gh-pages/paper</a>
**Version:** v0.1.0
