### 2021-05-24 #116 @pschloss
Possible reviewers. All suggested because of their prior interactions with the Carpentries and commitment to open science education...

* Tracy Teal
* Kim Dill-McFarland
* Kate Hertweck
