### 2021-06-11 cedee1a
Correct metadata for lua-filters
