### 2021-06-11 f4b2b57
whedon needs 'authors' not 'author'
while the lua filter requires 'author'
