### 2021-06-11 3fd122b
Remove old plain pandoc files
