### 2021-06-11 d35afbc
Add version with pandoc-scholar formatted header
