### 2021-06-15 42b4540
Add survey results figure
