### 2021-06-15 99727b3
Compile preprint with survey data figure
