### 2021-06-15 9f04f73
Update fig 3 caption
