### 2021-07-07 #116 @pschloss
Her GitHub handle is tracykteal

Pat

> On Jul 6, 2021, at 14:15, Shannon ***@***.***> wrote:
>
>
> Hi @pschloss <https://github.com/pschloss>, thanks for your patience and sorry for the delay. I was looking into tickets that could use a nudge and have sent out review invitations to Drs. Dill-McFarland and Hertweck (thank you for the names). I couldn't find contact information for Tracy Teal. Do you have this information?
>
