### 2021-07-30 #116 @magsol
@tracykteal Hi Tracy 👋 — would you consider contributing a peer review for JOSE? (The Journal of Open Source Education)
Submission title: "**Developing and deploying an integrated workshop curriculum teaching computational skills for reproducible research**"
