### 2021-10-12 #144 @whedon
[REVIEW]: Developing and deploying an integrated workshop curriculum teaching computational skills for reproducible research
**Submitting author:** @pschloss (<a href="http://orcid.org/0000-0002-6935-4275">Patrick D Schloss</a>)
**Repository:** <a href="https://github.com/UMCarpentries/intro-curriculum-r/" target ="_blank">https://github.com/UMCarpentries/intro-curriculum-r/</a>
**Version:** v1.0.0
