### 2021-10-15 f391043
Fix a few small bugs in 04-r-data-analysis.Rmd that prevent site from building
