+++ _episodes_rmd/04-r-data-analysis.Rmd
@@ -1,4 +1,4 @@
-  ---
+---
 title: "R for Data Analysis"
 source: Rmd
 teaching: 150
@@ -103,16 +103,16 @@ When we call `summarize()`, we can use any of the column names of our data objec
 
 We name this new column so we can use in a future argument. So the `averageLifeExp=` part tells `summarize()` to use "averageLifeExp" as the name of the new column. Note that you don't have to quotes around this new name as long as it starts with a letter and doesn't include a space.
 
-Instead of including the data as an argument, we can use the *pipe operator* `%>%` to pass the data value into the `summarize` function. 
+Instead of including the data as an argument, we can use the *pipe operator* `%>%` to pass the data value into the `summarize` function.
 
 ```{r AvgLifeExpWithPipe}
 gapminder_data %>% summarize(averageLifeExp=mean(lifeExp))
 ```
- 
+
 This line of code will do the exact same thing as our first summary command, but the piping function tells R to use the `gapminder_data` dataframe as the first argument in the next function.
 
 This lets us "chain" together multiple functions, which will be helpful later.
- 
+
 We can also add an <kdb>Enter</kdb> to make it look nicer:  
 ```{r AvgLifeExpWithPipe2}
 gapminder_data %>%
@@ -124,25 +124,25 @@ Using the *pipe operator* `%>%` and enter command makes our code more readable. 
 Since we use the pipe operator so often, there is a keyboard shortcut for it in RStudio. You can press <kdb>Ctrl</kdb>+<kdb>Shift</kdb>+<kdb>M<kdb> on Windows or <kdb>Cmd<kdb>+<kdb>Shift<kdb>+<kdb>M<kdb> on a Mac.
 
 > Helpful hint: Saving a new dataframe
-> Notice that when we run the following code, we are not actually saving a new variable: 
+> Notice that when we run the following code, we are not actually saving a new variable:
 > ```{r AvgLifeExpWithPipe3, eval = FALSE}
 > gapminder_data %>%
 >   summarize(averageLifeExp=mean(lifeExp))
 > ```
-> This simply outputs what we have created, but does not change actually change `gapminder_data` or save a new dataframe. To save a new dataframe, we could run: 
+> This simply outputs what we have created, but does not change actually change `gapminder_data` or save a new dataframe. To save a new dataframe, we could run:
 > ```{r saveNewDF, eval = FALSE}
 > gapminder_data_summarized <- gapminder_data %>%
 >   summarize(averageLifeExp=mean(lifeExp))
 > ```
-> Or if we want to change `gapminder_data` itself: 
+> Or if we want to change `gapminder_data` itself:
 > ```{r overwriteDF, eval = FALSE}
 > gapminder_data <- gapminder_data %>%
 >   summarize(averageLifeExp=mean(lifeExp))
 > ```
-> IMPORTANT: This would overwrite the existing `gapminder_data` object. 
-> 
-> For now, we will not be saving dataframes, since we are just experimenting with `dyplr` functions, but it will be useful later on in this lesson. 
-> 
+> IMPORTANT: This would overwrite the existing `gapminder_data` object.
+>
+> For now, we will not be saving dataframes, since we are just experimenting with `dyplr` functions, but it will be useful later on in this lesson.
+>
 {: .callout}
 
 ## Narrow down rows with `filter()` {#narrow-down-rows-with-filter}
@@ -165,7 +165,7 @@ Use `summarize()` to find the most recent year in the data set. We can use the `
 > > gapminder_data %>%  
 > >  summarize(recent_year = max(year))
 > > ```
-> > 
+> >
 > {: .solution}
 {: .challenge}
 
@@ -235,7 +235,7 @@ The `group_by()` function expects you to pass in the name of a column (or multip
 > {: .solution}
 {: .challenge}
[…]

[…]
