### 2021-10-15 7c9c4db
Added Zoom/Slack notes to welcome.
