+++ _episodes/00-welcome.md
@@ -2,30 +2,38 @@
 title: "Welcome to Software Carpentry"
 teaching: 15
 exercises: 0
-questions: 
+questions:
 - "What is The Carpentries?"
 - "What will the workshop cover?"
 - "What else do I need to know about the workshop?"
-objectives: 
+objectives:
 - "Introduce The Carpentries."
 - "Go over logistics."
 - "Introduce the workshop goals."
-keypoints: 
+keypoints:
 - "We follow The Carpentries Code of Conduct."
 - "Our goal is to generate a shareable and reproducible report by the end of the workshop."
 - "This lesson content is targeted to absolute beginners with no coding experience."
 ---
 
 
-## What is The Carpentries?
+## What is The Carpentries / UM Carpentries?
 
-The Carpentries is a global organization whose mission is to teach researchers, and others, the basics of coding so that you can use it in your own work. 
+The Carpentries is a global organization whose mission is to teach researchers, and others, the basics of coding so that you can use it in your own work.
 We believe everyone can learn to code, and that a lot of you will find it very useful for things such as data analysis and plotting.
 
-Our workshops are targeted to absolute beginners, and we expect that you have zero coding experience coming in. 
+Our workshops are targeted to absolute beginners, and we expect that you have zero coding experience coming in.
 That being said, you're welcome to attend a workshop if you already have a coding background but want to learn more!
 
-To provide an inclusive learning environment, we follow The Carpentries [Code of Conduct](https://docs.carpentries.org/topic_folders/policies/code-of-conduct.html). 
+The [Carpentries at the University of Michigan](#umcarpentries.org) is local
+group of data and software enthusiasts that regularly present workshops for
+UM groups and the community at large. Let us know if you have interest in
+hosting a workshop for your group, participating as a helper in a future workshop,
+or becoming trained as an instuctor.
+
+## Code of Conduct
+
+To provide an inclusive learning environment, we follow The Carpentries [Code of Conduct](https://docs.carpentries.org/topic_folders/policies/code-of-conduct.html).
 We expect that instructors, helpers, and learners abide by this code of conduct, including practicing the following behaviors:
 
 - Use welcoming and inclusive language.
@@ -40,17 +48,113 @@ You can report any violations to the Code of Conduct by filling out [this form](
 
 Now that you know a little about The Carpentries as an organization, the instructors and helpers will introduce themselves and what they'll be teaching/helping with.
 
-## The etherpad & introducing participants
+## Using Zoom and Slack
 
-Now it's time for the participants to introduce themselves. 
-Instead of verbally, the participants will use the _etherpad_ to write out their introduction. 
-We use the etherpad to take communal notes during the workshop. 
-Feel free to add your own notes on there whenever you'd like. 
-Go to the etherpad and write down your name, role, affiliation, and work/research area.
+### Arranging your screens
+It is important that you can see:
+
+ - Zoom (instructor's shared screen + reactions)
+ - Slack
+ - R/Studio or terminal/command window
+
+![screen setup]({{ page.root }}/fig/welcome/screen_setup.png)
+<br/>
+<br/>
+
+
+### Using Zoom
+
+- Zoom controls are at the bottom of the Zoom window:
+![Zoom controls]({{ page.root }}/fig/welcome/zoom_controls.png)
+
+- To minimize distractions, we encourage participants to keep their audio muted
+  (unless actively asking a question).
+- To maximize engagement, we encourage participants to keep their video on.
+- Slack works better than Zoom's Chat function so avoid Zoom Chat for now.
+- You can enable transcription subtitles for your view.
+- We will be using Breakout Rooms occasionally for ad-hoc 1-1 helper support.
[…]
