+++ _episodes/01-r-plotting.md
@@ -168,7 +168,7 @@ library(tidyverse)
 
 
 ~~~
-── Attaching packages ───────────────────────────────────────────────── tidyverse 1.3.1 ──
+── Attaching packages ─────────────────────────────────────── tidyverse 1.3.1 ──
 ~~~
 {: .output}
 
@@ -185,7 +185,7 @@ library(tidyverse)
 
 
 ~~~
-── Conflicts ──────────────────────────────────────────────────── tidyverse_conflicts() ──
+── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
 ✖ dplyr::filter() masks stats::filter()
 ✖ dplyr::lag()    masks stats::lag()
 ~~~
@@ -266,7 +266,7 @@ Rows: 142 Columns: 5
 
 
 ~~~
-── Column specification ──────────────────────────────────────────────────────────────────
+── Column specification ────────────────────────────────────────────────────────
 Delimiter: ","
 chr (2): country, continent
 dbl (3): pop, lifeExp, gdpPercap
@@ -392,7 +392,7 @@ Do all functions need arguments? Let's test some other functions:
 
 
 ~~~
-[1] "2021-10-14"
+[1] "2021-10-15"
 ~~~
 {: .output}
 
@@ -475,7 +475,7 @@ Rows: 142 Columns: 5
 
 
 ~~~
-── Column specification ──────────────────────────────────────────────────────────────────
+── Column specification ────────────────────────────────────────────────────────
 Delimiter: ","
 chr (2): country, continent
 dbl (3): pop, lifeExp, gdpPercap
@@ -844,7 +844,7 @@ Rows: 1704 Columns: 6
 
 
 ~~~
-── Column specification ──────────────────────────────────────────────────────────────────
+── Column specification ────────────────────────────────────────────────────────
 Delimiter: ","
 chr (2): country, continent
 dbl (4): year, pop, lifeExp, gdpPercap
@@ -1526,71 +1526,70 @@ animatedHansPlot
 
 
 ~~~
-Rendering [--------------------------------------------] at 3.2 fps ~ eta: 31s
-Rendering [>-------------------------------------------] at 3.1 fps ~ eta: 31s
-Rendering [=>------------------------------------------] at 3.1 fps ~ eta: 31s
-Rendering [==>-----------------------------------------] at 3.1 fps ~ eta: 31s
-Rendering [==>-----------------------------------------] at 3.1 fps ~ eta: 30s
-Rendering [===>------------------------------------------] at 3 fps ~ eta: 30s
-Rendering [====>-----------------------------------------] at 3 fps ~ eta: 30s
-Rendering [====>-----------------------------------------] at 3 fps ~ eta: 29s
-Rendering [=====>----------------------------------------] at 3 fps ~ eta: 29s
-Rendering [======>---------------------------------------] at 3 fps ~ eta: 28s
-Rendering [=======>--------------------------------------] at 3 fps ~ eta: 28s
-Rendering [=======>--------------------------------------] at 3 fps ~ eta: 27s
-Rendering [========>-------------------------------------] at 3 fps ~ eta: 27s
-Rendering [=========>------------------------------------] at 3 fps ~ eta: 27s
-Rendering [=========>------------------------------------] at 3 fps ~ eta: 26s
-Rendering [==========>-----------------------------------] at 3 fps ~ eta: 26s
-Rendering [===========>----------------------------------] at 3 fps ~ eta: 25s
-Rendering [===========>--------------------------------] at 2.9 fps ~ eta: 25s
-Rendering [============>-------------------------------] at 2.9 fps ~ eta: 24s
-Rendering [=============>------------------------------] at 2.9 fps ~ eta: 24s
[…]

[…]
