+++ _episodes/03-intro-git-github.md
@@ -1439,6 +1439,8 @@ Now that we have two repositories, we need a diagram like this:
 Note that our local repository still contains our earlier work on `notes.txt`, but the
 remote repository on GitHub appears empty as it doesn't contain any files yet.
 
+### Linking a local repository to GitHub
+
 The next step is to connect the two repositories.  We do this by making the
 GitHub repository a [remote]({{ page.root}}{% link reference.md %}#remote) for the local repository.
 The home page of the repository on GitHub includes the string we need to
@@ -1474,13 +1476,48 @@ origin   https://github.com/USERNAME/un-report.git (fetch)
 ```
 {: .output}
 
-Once the remote is set up, this command will push the changes from
-our local repository to the repository on GitHub:
+Now we want to send our local git information to GitHub. While the default for code you put on GitHub is that anyone can view or make copies of your code, in order to make changes to your repository, you need to be able to log in so GitHub can recognize you as someone who is authorized to make changes.
+
+### Setting up your GitHub Personal Access Token (PAT)
+
+When you use the GitHub website, you need to login with a username and password. By default, only you will be able to make any changes to the repositories you create. In order to perform git commands on your own computer that interact with GitHub, we need a way to tell GitHub who you are. Rather than requiring you to type your password every time, you can create identify yourself with a personal access token (PAT). Let's first tell git that we would like it to remember our credentials so we don't have to constantly retype them. At the command line type:
+
+```
+git config --global credential.helper store
+```
+{: .language-bash}
+
+Like the previous `git config` commands we ran before, this tells git to store our account information so it doesn’t have to ask us for it every time we use git on the command line.
+
+The information git stores is your personal access token (PAT). These tokens are basically a secret word that only you know that allows you to access all your stuff. Think of these tokens like a key to your house. You never want to hand over the keys to your house to someone you don't trust. But as long as you hang on to that key, you are free to access all your stuff.
+
+> ## What's the difference between passwords and PATs?
+> You might be wondering why we can't just type a password to login and need to use a PAT instead. There are a few reasons:
+> * Human created passwords maybe be easy to guess and are often reused across many sites. You don't want to make it easy for someone to copy your keys nor is it safe just to have one key that can unlock everything you own (your house, your car, your secret money vault, etc)
+> * PATs are generated by computers, for computers. The PATs are much longer than most human created passwords and have random combinations of letters and characters that are very difficult to guess
+> * A user can generate multiple PATs for the same account for different uses with different permissions
+> * Github now requires the use of PATs when using HTTPS (so we don't really have a choice)
+> Overall PATs are more secure if you also keep them private.
+{: .callout}
+
+To create a PAT, you'll need to be logged into to GitHub. Click your profile icon on the top left and choose "Setting" from the dropdown. On the main settings page there is a long list of options on the left. Scroll down till you find "Developer Settings". Next you should see three options: "GitHub Apps", "OAuth Apps", and "Personal access tokens". We want to create a token so click on the last link. You should now see a link to " Generate a personal access token". Click that. (You should now be at https://github.com/settings/tokens/new)
+
+On the "New personal access token" form, the first field you see is for "Note." You can actually create multiple tokens. This note field helps you remember what the token was for. It's a good idea to create one per computer you use so the note field would be something like "work-laptop", "home-macbook", or "greatlakes-project". Next you will see an option for "Expiration." Since your tokens are like the keys to your house, it might be a good idea that if you forgot about your tokens, they just stop working after a while so no one else can misuse them. When your tokens expire, you can just generate a new one. GitHub recommends you choose an expiration date so we can just choose "90 days" or whatever is appropriate for you. (Note: You will have to repeat this processes of generating a new PAT when an existing PAT expires.)
+
[…]
