### 2021-11-10 UMCarpentries/intro-curriculum-r#147 @kdillmcfarland
Improve Linux installation 
The Linus installation section in setup.html is incomplete. Given that the target audience is novice coders, it should be expanded to include all the same sections as Mac and Windows e.g.

- The Unix Shell
- Installing Git
- Installing R
- Installing RStudio
