### 2021-11-10 UMCarpentries/intro-curriculum-r#148 @kdillmcfarland
Expand index page introduction
As in original Carpentries lessons, it would be helpful to have a summary of the module, major learning objects, and target audience described on the index page prior to viewing individual lessons. This information is in the `1. Welcome to Software Carpentry` lesson but that title does not make it clear what is covered there.
