### 2021-11-10 #144 @kdillmcfarland
Documentation

- Request for improved Linux installation https://github.com/UMCarpentries/intro-curriculum-r/issues/147
- Request for expanded statement of purpose https://github.com/UMCarpentries/intro-curriculum-r/issues/148
