### 2021-11-10 97a314f
Typos and other minor corrections
As part of JOSE review
