+++ _episodes_rmd/01-r-plotting.Rmd
@@ -239,7 +239,7 @@ You should now have a line of text in your code file that started with `gapminde
 
 What if we want to run this command from our code file?
 
-In order to run code that you've typed in the editor, you have a few options. We can click <kbd>Run</kbd> again from the right side of the **Editor** tab but the quickest way to run the code is by pressing <kbd>Ctrl</kbd>+<kbd>Enter</kbd> on your keyboard.
+In order to run code that you've typed in the editor, you have a few options. We can click <kbd>Run</kbd> again from the right side of the **Editor** tab but the quickest way to run the code is by pressing <kbd>Ctrl</kbd>+<kbd>Enter</kbd> on your keyboard (<kbd>Ctrl</kbd>+<kbd>Enter</kbd> on Mac).
 
 This will run the line of code that currently contains your cursor and will move your cursor to the next line. Note that when Rstudio runs your code, it basically just copies your code from the **Editor** window to the **Console** window, just like what happened when we selected <kbd>Run Selected Line(s)</kbd>.
 
@@ -535,7 +535,7 @@ There are also lots of other fun options:
 > {: .solution}
 {: .challenge}
 
-Since we have the data for the population of each country, we might be curious what effect population might have on life expectancy and GDP per capita. Do you think larger countires will have a longer or shorter life expectancy? Let's find out by mapping the population of each country to the size of our points.
+Since we have the data for the population of each country, we might be curious what effect population might have on life expectancy and GDP per capita. Do you think larger countries will have a longer or shorter life expectancy? Let's find out by mapping the population of each country to the size of our points.
 
 ```{r FirstPlotAddSize}
 ggplot(data = gapminder_1997) +
@@ -613,7 +613,7 @@ Many datasets are much more complex than the example we used for the first plot.
 ## Importing datasets
 _[Back to top](#contents)_
 
-In the first plot, we looked at a smaller slice of a large dataset. To gain a better understanding of the kinds of patterns we might observe in our own data, we will now use the full dataset, which is stored in a called "gapminder_data.csv".
+In the first plot, we looked at a smaller slice of a large dataset. To gain a better understanding of the kinds of patterns we might observe in our own data, we will now use the full dataset, which is stored in a file called "gapminder_data.csv".
 
 To start, we will read in the data without using the interactive RStudio file navigation.
 
@@ -707,7 +707,7 @@ Sometimes plots like this are called "spaghetti plots" because all the lines loo
 > >  aes(x = pop, y = lifeExp, group = country, color = continent) +
 > >  geom_line()
 > > ```
-> > (China and India are the two Asian countries that have experience massive population growth from 1952-2007.)
+> > (China and India are the two Asian countries that have experienced massive population growth from 1952-2007.)
 > {: .solution}
 {: .challenge}
 
@@ -732,8 +732,8 @@ We've previously used the discrete values of the `continent` column to color in 
 
 This type of visualization makes it easy to compare the range and spread of values across groups. The "middle" 50% of the data is located inside the box and outliers that are far away from the central mass of the data are drawn as points.
 
-> ## Bonus Exercise:
-> Take a look a the ggplot cheat sheet. Find all the geoms listed under "Discrete X, Continuous Y". Try replacing `geom_boxplot` with one of these other functions.
+> ## Bonus Exercise: Other discrete geoms
+> Take a look a the ggplot [cheat sheet](https://ggplot2.tidyverse.org/). Find all the geoms listed under "Discrete X, Continuous Y". Try replacing `geom_boxplot` with one of these other functions.
 >
 > > ## Example solution
 > > ```{r GapViol}
@@ -903,7 +903,7 @@ ggplot(gapminder_1997) +
[…]
