+++ index.md
@@ -6,11 +6,19 @@ permalink: index.html  # Is the only page that doesn't follow the pattern /:path
 
 {% include gh_variables.html %}
 
-This is a custom curriculum developed by the University of Michigan Carpentries instance. It is in the `alpha` stage of development. 
+This is a custom curriculum developed by the University of Michigan 
+Carpentries instance. It is in the `alpha` stage of development. 
 
-The curriculum is developed for complete beginners interested in learning reproducible data science techniques and integrates instruction of R for data cleaning, analysis, and visualization; the Unix shell; and git and GitHub. For more details, please see [Introduction to the Workshop](https://umcarpentries.org/intro-curriculum-r/00-welcome/index.html). 
+The curriculum is developed for complete beginners interested in 
+learning reproducible data science techniques and integrates instruction 
+of R for data cleaning, analysis, and visualization; the Unix shell; and 
+git and GitHub. For more details, please see 
+[Introduction to the Workshop](https://umcarpentries.org/intro-curriculum-r/00-welcome/index.html). 
 
-If you are interested in contributing to the curriculum, please read over [The Carpentries Curriculum Development Handbook](https://carpentries.github.io/curriculum-development/) and [our contribution guidelines][contributing]. If you would like to set up your machine to preview changes locally,
+If you are interested in contributing to the curriculum, please read over 
+[The Carpentries Curriculum Development Handbook](https://carpentries.github.io/curriculum-development/) 
+and [our contribution guidelines][contributing]. If you would like to set 
+up your machine to preview changes locally,
 please see [the setup instructions]({{ page.root }}{% link setup.md %}).
 
 {% include links.md %}
