### 2021-11-22 d9b9a21
Udpate install instructions resolves #147
