### 2021-11-22 4e0df60
Add connection between unix flags and R function arguments
