### 2021-11-22 e5c7ea2
Create shortcut to setup instructions
