### 2021-11-22 f72564f
Use '{{ page.root }}/01-r-plotting/' syntax to link to internal lessons
