+++ _episodes_rmd/01-r-plotting.Rmd
@@ -24,7 +24,7 @@ keypoints:
 ---
 
 <!-- Begin Link shortcuts -->
-[setup_instructions]: http://umcarpentries.github.io/intro-curriculum-r/setup.html
+[setup_instructions]: {{ page.root }}/setup/
 <!-- End Link shortuts -->
 
 ```{r, include=FALSE}
@@ -174,7 +174,7 @@ library(tidyverse)
 
 > ## What's with all those messages???
 >
-> When you loaded the `tidyverse` package, you probably got a message like the one we got above. Don't panic! These messages are just giving you more information about what happened when you loaded `tidyverse`. `Tidyverse` is actually a collection of several different packages, so the first section of the message tells us what packages were installed when we loaded `tidyverse` (these include `ggplot2`, which we'll be using a lot in this lesson, and `dyplr`, which you'll be introduced to tomorrow in the [R for Data Analysis lesson](05-r-markdown.Rmd)).
+> When you loaded the `tidyverse` package, you probably got a message like the one we got above. Don't panic! These messages are just giving you more information about what happened when you loaded `tidyverse`. `Tidyverse` is actually a collection of several different packages, so the first section of the message tells us what packages were installed when we loaded `tidyverse` (these include `ggplot2`, which we'll be using a lot in this lesson, and `dyplr`, which you'll be introduced to tomorrow in the [R for Data Analysis lesson]({{ page.root }}/04-r-data-analysis/)).
 
 > The second section of messages gives a list of "conflicts." Sometimes, the same function name will be used in two different packages, and R has to decide which function to use. For example, our message says that:
 >
