+++ _episodes/02-unix-shell.md
@@ -159,18 +159,13 @@ This error message tells us the command we tried to run, `ks`, is not a command 
 ## Man and Help
 _[Back to top](#contents)_
 
-Now that we know how to list files with `ls`, we can learn how to look up the manual pages for unix shell commands. If you want to learn more about a command, we can use `man` to look up its manual page. We can navigate the man page to view the description of a command and its options. For example, if you want to know more about the navigation options of `ls` you can type `man ls` on the command line.
+Often we'll want to learn more about how to use a certain command such as `ls`. There are several different ways you can 
+learn more about a specific command. 
 
-```
-man ls
-```
-{: .language-bash}
+Some commands have additional information that can be found by using the `-h` or `--help`
+flags. This will print brief documentation for the command:
 
 ![]({{ page.root }}/fig/unix-shell/man_ls.png)
-
-On the manual page for `ls`, we see a section titled **options**. These options, also called **flags**, are similar to arguments in R functions, and allow us to customize how `ls` runs.
-
-One very helpful flag that is available for any command is `-h or --help` which will print brief documentation for the command.
 
 ```
 man -h
@@ -180,6 +175,32 @@ man --help
 
 ![]({{ page.root }}/fig/unix-shell/man_help.png)
 
+Other commands, such as `ls`, don't have help flags, but have manual pages with more information. We can navigate 
+the manual page using the `man` command to view the description of a command and its options. For example, 
+if you want to know more about the navigation options of `ls` you can type `man ls` on the command line:
+
+```
+man ls
+```
+{: .language-bash}
+
+On the manual page for `ls`, we see a section titled **options**. These options, also called **flags**, are similar to arguments in R functions, and allow us to customize how `ls` runs.
+
+To get out of the man page, click `q`.
+
+Sometimes, commands will have multiple flags that we want to use at the same time. For example, 
+`ls` has a flag `-F` that displays a slash after all directories, as well as a flag `-a` that
+includes hidden files and directories (ones that begin with a `.`). There are two ways to run 
+`ls` using both of these flags:
+
+```
+ls -F -a
+ls -Fa
+```
+{: .language-bash}
+
+Note that when we run the `-a` command, we see a `.` and a `..` in the directory. The `.` corresponds to the current directory we are in and the `..` corresponds to the directory directly above us in the directory tree. We'll learn more about why this is useful in a bit.
+
 > ## Using the Manual Pages
 >
 > Use `man` to open the manual for the command `ls`.
@@ -187,18 +208,16 @@ man --help
 > What flags would you use to...
 > 1. Print files in order of size?
 > 2. Print files in order of the last time they were edited?
-> 3. Print hidden files (files that begin with `.`)?
-> 4. Print more information about the files?
-> 5. Print more information about the files with unit suffixes?
-> 6. Print files in order of size AND also print more information about the files?
+> 3. Print more information about the files?
+> 4. Print more information about the files with unit suffixes?
+> 5. Print files in order of size AND also print more information about the files?
 >
 > > ## Solution
 > > 1. `ls -S`
 > > 2. `ls -t`
-> > 3. `ls -a`
-> > 4. `ls -l`
-> > 5. `ls -lh`
-> > 6. `ls -lS`
+> > 3. `ls -l`
+> > 4. `ls -lh`
+> > 5. `ls -lS`
 >
 > {: .solution}
 {: .challenge}
@@ -640,32 +659,38 @@ _[Back to top](#contents)_
 
 To view and navigate the contents of a file we can use the command `less`. This will open a full screen view of the file.
 
-Here is what we should expect to see when running the command `less` on our `gapminder_data.csv` file:
+For instance, we can run the command `less` on our `gapminder_data.csv` file:
+
+```
[…]
