+++ _episodes_rmd/01-r-plotting.Rmd
@@ -328,13 +328,6 @@ Do all functions need arguments? Let's test some other functions:
   getwd()
 ```
 
-> ## Comments
-> Sometimes you may want to write comments in your code to help you remember what your code is doing, but you don't want R to think these comments are a part of the code you want to evaluate. That's where **comments** come in! Anything after a `#` symbol in your code will be ignored by R. For example, let's say we wanted to make a note of what each of the functions we just used do:
-> ```{r FunctionsComments}
->  Sys.Date()  # outputs the current date
->  getwd()     # outputs our current working directory (folder)
->```
-{: .callout}
 
 While some functions, like those above, don't need any arguments, in other functions we may want to use multiple arguments. When we're using multiple arguments, we separate the arguments with commas. For example, we can use the `sum()` function to add numbers together:
 
@@ -392,6 +385,17 @@ Sometimes it is helpful - or even necessary - to include the argument name, but 
 > > One way is using the `read_excel` function in the `readxl` package. There are other ways, but this is our preferred method because the output will be the same as the output of `read_csv`.
 > {: .solution}
 {: .challenge}
+
+> ## Comments
+> Sometimes you may want to write comments in your code to help you remember what your code is doing, but you don't want R to think these comments are a part of the code you want to evaluate. That's where **comments** come in! Anything after a `#` symbol in your code will be ignored by R. For example, let's say we wanted to make a note of what each of the functions we just used do:
+> ```{r FunctionsComments}
+>  Sys.Date()  # outputs the current date
+>  getwd()     # outputs our current working directory (folder)
+>  sum(5, 6)   # adds numbers
+>  read_csv(file = 'gapminder_1997.csv') # reads in csv file
+>```
+{: .callout}
+
 
 # Creating our first plot
 _[Back to top](#contents)_
