### 2021-11-22 7e70d09
Show plots in plotting for data exploration
