### 2021-11-22 c291257
Fix gh-pages typos
