### 2021-11-22 363d72a
Add note that we haven not learned about data cleaning yet but will in next lesson
