### 2021-11-22 e3da2fd
Remove blue color from bonus topics
