### 2021-11-22 4d0c186
Add tidyr to learning objectives
