### 2021-11-22 e0c0ea0
Remove blank space before header
