### 2021-11-22 5f91a1f
Move join figures and update paths
