### 2021-11-22 00f97a1
Explain difference between pipe and plus
