### 2021-11-22 ecf128d
Fix link to setup page for real this time
