### 2021-11-22 dec3f6d
Remove example R markdown images
