+++ _episodes/01-r-plotting.md
@@ -1,0 +1,1695 @@
+---
+# Please do not edit this file directly; it is auto generated.
+# Instead, please edit 01-r-plotting.md in _episodes_rmd/
+source: Rmd
+title: "R for Plotting"
+teaching: 120
+exercises: 30
+questions:
+- "What are R and R Studio?"
+- "How do I read data into R?"
+- "What are geometries and aesthetics?"
+- "How can I use R to create and save professional data visualizations?"
+objectives:
+- "To become oriented with R and R Studio."
+- "To be able to read in data from csv files."
+- "To create plots with both discrete and continuous variables."
+- "To understand mapping and layering using `ggplot2`."
+- "To be able to modify a plot's color, theme, and axis labels."
+- "To be able to save plots to a local directory."
+keypoints:
+- "R is a free programming language used by many for reproducible data analysis."
+- "Use `read_csv` to read tabular data in R."
+- "Geometries are the visual elements drawn on data visualizations (lines, points, etc.), and aesthetics are the visual properties of those geometries (color, position, etc.)."
+- "Use `ggplot()` and geoms to create data visualizations, and save them using `ggsave()`."
+
+---
+
+
+
+
+
+
+### Contents
+1. [Introduction to R and RStudio](#introduction-to-r-and-rstudio)
+1. [Loading and reviewing data](#loading-and-reviewing-data)
+1. [Understanding commands](#understanding-commands)
+1. [Creating our first plot](#creating-our-first-plot)
+1. [Plotting for data exploration](#plotting-for-data-exploration)
+    + [Importing datasets](#importing-datasets)
+    + [Discrete plots](#discrete-plots)
+    + [Layers](#layers)
+    + [Color vs. Fill](#color-vs.-fill)
+    + [Univariate plots](#univariate-plots)
+    + [Plot themes](#plot-themes)
+    + [Facets](#facets)
+    + [Saving plots](#saving-plots)
+1. [Bonus topics!] 
+    + [Creating complex plots](#bonus-topics:-creating-complex-plots)
+      + [Animated plots](#animated-plots)
+      + [Map plots](#map-plots)
+1. [Glossary of terms](#glossary-of-terms)
+
+> ## Bonus: why learn to program?
+> Share why you're interested in learning how to code.
+> > ## Solution:
+> > There are lots of different reasons, including to perform data analysis and generate figures. I'm sure you have morespecific reasons for why you'd like to learn!
+> {: .solution}
+{: challenge}
+
+# Introduction to R and RStudio
+_[Back to top](#contents)_
+
+In this session we will be testing the hypothesis that a country's life expectancy is related to the total value of its finished goods and services, also known as the Gross Domestic Product (GDP).
+To test this hypothesis, we'll need two things: data and a platform to analyze the data.
+
+You already [downloaded the data]({{ page.root }}/setup/). But what platform will we use to analyze the data? We have many options!
+
+We could try to use a spreadsheet program like Microsoft Excel or Google sheets that have limited access, less flexibility, and don't easily allow for things that are critical to ["reproducible" research](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1003285), like easily sharing the steps used to explore and make changes to the original data.
+
+Instead, we'll use a programming language to test our hypothesis. Today we will use R, but we could have also used Python for the same reasons we chose R (and we teach workshops for both languages). Both R and Python are freely available, the instructions you use to do the analysis are easily shared, and by using reproducible practices, it's straightforward to add more data or to change settings like colors or the size of a plotting symbol.
+
+> ## But why R and not Python?
+> [There's no great reason](http://www.academichermit.com/2020/03/23/Why-R.html).
[…]
