+++ _episodes/01-r-plotting.md
@@ -169,7 +169,7 @@ library(tidyverse)
 
 
 ~~~
-── Attaching packages ───────────────────────────────────────────────────────────────────────── tidyverse 1.3.1 ──
+── Attaching packages ─────────────────────────────────────── tidyverse 1.3.1 ──
 ~~~
 {: .output}
 
@@ -186,7 +186,7 @@ library(tidyverse)
 
 
 ~~~
-── Conflicts ──────────────────────────────────────────────────────────────────────────── tidyverse_conflicts() ──
+── Conflicts ────────────────────────────────────────── tidyverse_conflicts() ──
 ✖ dplyr::filter() masks stats::filter()
 ✖ dplyr::lag()    masks stats::lag()
 ~~~
@@ -267,7 +267,7 @@ Rows: 142 Columns: 5
 
 
 ~~~
-── Column specification ──────────────────────────────────────────────────────────────────────────────────────────
+── Column specification ────────────────────────────────────────────────────────
 Delimiter: ","
 chr (2): country, continent
 dbl (3): pop, lifeExp, gdpPercap
@@ -407,7 +407,7 @@ Do all functions need arguments? Let's test some other functions:
 
 
 ~~~
-[1] "/Users/kelly/projects/carpentries/intro-curriculum-r/_episodes_rmd"
+[1] "/home/runner/work/intro-curriculum-r/intro-curriculum-r/_episodes_rmd"
 ~~~
 {: .output}
 
@@ -436,7 +436,7 @@ Do all functions need arguments? Let's test some other functions:
 > 
 > 
 > ~~~
-> [1] "/Users/kelly/projects/carpentries/intro-curriculum-r/_episodes_rmd"
+> [1] "/home/runner/work/intro-curriculum-r/intro-curriculum-r/_episodes_rmd"
 > ~~~
 > {: .output}
 {: .callout}
@@ -488,7 +488,7 @@ Rows: 142 Columns: 5
 
 
 ~~~
-── Column specification ──────────────────────────────────────────────────────────────────────────────────────────
+── Column specification ────────────────────────────────────────────────────────
 Delimiter: ","
 chr (2): country, continent
 dbl (3): pop, lifeExp, gdpPercap
@@ -857,7 +857,7 @@ Rows: 1704 Columns: 6
 
 
 ~~~
-── Column specification ──────────────────────────────────────────────────────────────────────────────────────────
+── Column specification ────────────────────────────────────────────────────────
 Delimiter: ","
 chr (2): country, continent
 dbl (4): year, pop, lifeExp, gdpPercap
@@ -1532,92 +1532,6 @@ animatedHansPlot <- staticHansPlot +
 animatedHansPlot
 ~~~
 {: .language-r}
-
-
-
-~~~
-Rendering [--------------------------------------------] at 1.3 fps ~ eta: 1m
-Rendering [>-------------------------------------------] at 1.3 fps ~ eta: 1m
-Rendering [>-------------------------------------------] at 1.1 fps ~ eta: 1m
-Rendering [=>--------------------------------------------] at 1 fps ~ eta: 2m
-Rendering [==>-------------------------------------------] at 1 fps ~ eta: 2m
-Rendering [==>-----------------------------------------] at 1.1 fps ~ eta: 1m
-Rendering [===>----------------------------------------] at 1.1 fps ~ eta: 1m
-Rendering [====>---------------------------------------] at 1.1 fps ~ eta: 1m
-Rendering [=====>--------------------------------------] at 1.1 fps ~ eta: 1m
-Rendering [======>-------------------------------------] at 1.1 fps ~ eta: 1m
-Rendering [======>-------------------------------------] at 1.2 fps ~ eta: 1m
[…]
