### 2021-11-22 dc22664
Add mention of R Studio terminal
