+++ _episodes/03-intro-git-github.md
@@ -350,7 +350,7 @@ $ git status
 ```
 {: .language-bash}
 ```
-On branch master
+On branch main
 
 No commits yet
 
@@ -505,7 +505,7 @@ $ git status
 {: .language-bash}
 
 ```
-On branch master
+On branch main
 
 No commits yet
 
@@ -535,7 +535,7 @@ $ git status
 {: .language-bash}
 
 ```
-On branch master
+On branch main
 
 No commits yet
 
@@ -558,7 +558,7 @@ $ git commit -m "Start notes on analysis"
 {: .language-bash}
 
 ```
-[master (root-commit) f22b25e] Start notes on analysis
+[main (root-commit) f22b25e] Start notes on analysis
  1 file changed, 1 insertion(+)
  create mode 100644 notes.txt
 ```
@@ -588,7 +588,7 @@ $ git status
 {: .language-bash}
 
 ```
-On branch master
+On branch main
 nothing to commit, working directory clean
 ```
 {: .output}
@@ -654,7 +654,7 @@ $ git status
 {: .language-bash}
 
 ```
-On branch master
+On branch main
 Changes not staged for commit:
   (use "git add <file>..." to update what will be committed)
   (use "git checkout -- <file>..." to discard changes in working directory)
@@ -717,7 +717,7 @@ $ git status
 {: .language-bash}
 
 ```
-On branch master
+On branch main
 Changes not staged for commit:
   (use "git add <file>..." to update what will be committed)
   (use "git checkout -- <file>..." to discard changes in working directory)
@@ -739,7 +739,7 @@ $ git commit -m "Add information on points"
 {: .language-bash}
 
 ```
-[master 34961b1] Add information on points
+[main 34961b1] Add information on points
  1 file changed, 1 insertion(+)
 ```
 {: .output}
@@ -867,7 +867,7 @@ $ git commit -m "Add note about point color"
 {: .language-bash}
 
 ```
-[master 005937f] Add note about point color
+[main 005937f] Add note about point color
  1 file changed, 1 insertion(+)
 ```
 {: .output}
@@ -880,7 +880,7 @@ $ git status
 {: .language-bash}
 
 ```
-On branch master
+On branch main
 nothing to commit, working directory clean
 ```
 {: .output}
@@ -974,7 +974,7 @@ Date:   Thu Aug 22 09:51:46 2020 -0400
 > You can also combine the `--oneline` option with others. One useful
 > combination adds `--graph` to display the commit history as a text-based
 > graph and to indicate which commits are associated with the
-> current `HEAD`, the current branch `master`, or
+> current `HEAD`, the current branch `main`, or
 > [other Git references][git-references]:
 >
 > ```
@@ -982,7 +982,7 @@ Date:   Thu Aug 22 09:51:46 2020 -0400
 > ```
 > {: .language-bash}
 > ```
-> * 005937f (HEAD -> master) Add note about point color
+> * 005937f (HEAD -> main) Add note about point color
 > * 34961b1 Add information on points
 > * f22b25e Start notes on analysis
 > ```
@@ -1147,7 +1147,7 @@ repository (`git commit`):
 > > ```
 > > {: .language-bash}
 > > ```
-> > [master cc127c2]
+> > [main cc127c2]
 > >  Note plans to start a draft manuscript
 > >  2 files changed, 2 insertions(+)
 > >  create mode 100644 manuscript.txt
@@ -1388,7 +1388,7 @@ Systems like Git allow us to move work between any two repositories. In
 practice, though, it's easiest to use one copy as a central hub, and to keep it
 on the web rather than on someone's laptop. Most programmers use hosting
 services like [GitHub](https://github.com), [Bitbucket](https://bitbucket.org) or
-[GitLab](https://gitlab.com/) to hold those master copies.
[…]
