### 2021-11-22 c30e34e
Change git steps to numbered list
