### 2021-11-22 853dbac
Explain that rendered kable will look similar to what it looks like on the website
