+++ _episodes_rmd/05-r-markdown.Rmd
@@ -372,12 +372,14 @@ If you don't already have your partner's GitHub repo cloned from the git/GitHub 
 The way you will collaborate with each other is as follows:
 1. For each exercise, both people will be thinking about how to answer the question, but only one person will be writing the code.
 This is called _paired programming_.
-1. Once you have completed an exercise, the person working on the exercise will add, commit, and push the changes to GitHub.
+1. Once you have completed 3 exercises, the person working on the exercises will add, commit, and push the changes to GitHub.
 1. Then the other person will pull the changes from GitHub.
 1. The person who pulled changes will code for the next exercise.
 1. Repeat the process for as many exercises as you can finish in the remaining time.
+
 Don't worry if you don't finish all of the exercises, and it's not a race between groups!
 This is just a way for you to practice what you've learned.
+Also, you can switch off more or less frequently depending on how much you want to practice pushing and pulling to/from GitHub.
 
 **One note:** It may be helpful to copy and paste the questions into the R Markdown file as you go.
 
@@ -386,7 +388,7 @@ _[Back to top](#contents)_
 
 First we're going to start out with a few questions about the gapminder dataset.
 
-1. The very first step is to read in the gapminder dataset, so don't forget! Also load the `tidyverse` package.
+[1] The very first step is to read in the gapminder dataset, so do that first! Also load the `tidyverse` package.
 
 > ## Solution
 > ```{r}
@@ -398,7 +400,7 @@ First we're going to start out with a few questions about the gapminder dataset.
 #### Investigating population over time.
 _[Back to top](#contents)_
 
-2. Make a scatter plot of year vs. population, separated into a plot for each contient. **HINT:** you can use `facet_wrap(vars(column_name))` to separate into different plots based on that column.
+[2] Make a scatter plot of year vs. population, separated into a plot for each contient. **Hint:** you can use `facet_wrap(vars(column_name))` to separate into different plots based on that column.
 
 > ## Solution
 > ```{r}
@@ -408,15 +410,15 @@ _[Back to top](#contents)_
 > ```
 {: .solution}
 
-3. It seems like there are 2 outliers - which countries are those?
+[3] It seems like there are 2 outliers - which countries are those?
 
 > ## Solution
->```{r}
+> ```{r}
 > gapminder %>% filter(pop > 1e9) %>% select(country) %>% unique()
 > ```
 {: .solution}
 
-4. Plot year vs. population separated into a plot for each continent but excluding the 2 outlier countries.
+[4] Plot year vs. population separated into a plot for each continent but excluding the 2 outlier countries.
 
 > ## Solution
 > ```{r}
@@ -430,7 +432,7 @@ _[Back to top](#contents)_
 ##### Bonus questions: come back to these if you have time at the end
 _[Back to top](#contents)_
 
-5. In the plot above, the years look kind of messy. Can you rotate the x axis text 90 degrees so that the years are more readable? Feel free to search the internet if you don't know how to do this!
+[5] In the plot above, the years look kind of messy. Can you rotate the x axis text 90 degrees so that the years are more readable? Feel free to search the internet if you don't know how to do this!
 
 > Solution
 > ```{r}
@@ -441,7 +443,7 @@ _[Back to top](#contents)_
 > ```
 {: .solution}
 
-6. It's hard to see which country is which here. Can you change the scatter plot to a line plot so we can get a better sense of trends over time? HINT: This website has more information: https://www.r-graph-gallery.com/line-chart-several-groups-ggplot2.html
[…]
