### 2021-11-22 c318f9e
Move 'undoing changes' to after 'exploring history' in bonus section
