+++ .github/workflows/website.yml
@@ -17,6 +17,11 @@ jobs:
       run:
         shell: bash
     steps:
+      - uses: actions/checkout@v2
+        with:
+          persist-credentials: false
+          fetch-depth: 0
+
       - name: Set up Ruby
         uses: ruby/setup-ruby@v1
         with:
@@ -35,12 +40,6 @@ jobs:
       - name: Install Python modules
         run: |
           python3 -m pip install --upgrade pip setuptools wheel pyyaml==5.3.1 requests
-
-      - name: Checkout the lesson
-        uses: actions/checkout@master
-        with:
-          fetch-depth: 0
-          ref: ${{ github.event.pull_request.head.sha }}
 
       - name: Look for R-markdown files
         id: check-rmd
@@ -78,7 +77,6 @@ jobs:
           saveRDS(remotes::dev_package_deps(dependencies = TRUE), ".github/depends.Rds", version = 2)
           writeLines(sprintf("R-%i.%i", getRversion()$major, getRversion()$minor), ".github/R-version")
         shell: Rscript {0}
-
 
       - name: Install system dependencies for R packages
         if: steps.check-rmd.outputs.count != 0
