### 2021-11-22 1e37343
Add solutions to bonus exercises
