### 2021-11-22 ed0d10f
Just run the workflow on PRs
