### 2021-11-22 9ac6a3c
Throw an error if there are bad links
