### 2021-11-22 beb6cac
Run `make site` manually
WHY ISN'T THE GITHUB ACTIONS WORKFLOW WORKING?! 🤬
