+++ _episodes/05-r-markdown.md
@@ -48,26 +48,15 @@ Recall that our  goal is to generate a report to the United Nations on how a cou
 ## What is R Markdown and why use it?
 _[Back to top](#contents)_
 
-We're going to learn a way to create a report in R using something called R Markdown. R Markdown combines Markdown (which renders plain text) with R code, which you can use to do data analysis and make figures. R Markdown enables you to "render" a report of our analysis and results into an html, PDF or Word file that you can share with others. 
+In R Markdown, you can incorporate ordinary text (ex. experimental methods, analysis and discussion of results) alongside code and figures! (Some people write entire manuscripts in R Markdown.) This is useful for writing reproducible reports and publications, sharing work with collaborators, writing up homework, and keeping a bioinformatics notebook. Because the code is emedded in the document, the tables and figures are *reproducible*. Anyone can run the code and get the same results. If you find an error or want to add more to the report, you can just re-run the document and you'll have updated tables and figures! This concept of combining text and code is called "literate programming". To do this we use R Markdown, which combines Markdown (renders plain text) with R. You can output an html, PDF, or Word document that you can share with others. In fact, this webpage is an example of a rendered R markdown file!
 
-Combining Markdown and R code in the same document allows you to integrate the text of the report, the code you used, and your figures, tables, and other output into one document. The text can be your introduction, description of the research design and  experimental methods, your analysis and discussion of the results, the conclusion, or whatever else you want to include. R Markdown reports can be used for many different things, including reproducible analysis reports, publication, sharing work with collaborators, writing up homework, or keeping a computational notebook of your work. (If you are familiar with Jupyter notebooks in the Python programming environment, R Markdown is  R's equivalent of a Jupyter notebook.)
-
-![]({{ page.root }}/fig/r-markdown/report-no-code.png)
-![]({{ page.root }}/fig/r-markdown/report-with-code.png)
-![]({{ page.root }}/fig/r-markdown/r-markdown-code-only.png)
-
-
-
-
-We call  reports *reproducible*  when the code is embedded in the document and anyone can take the report and re-run the code to get the same output. If you're writing a publication or sharing results with collaborators who won't be interested in seeing that nitty-gritty part of your work, you can  hide the code in the final, rendered output file while keeping it in the R Markdown file for your own record.
-Writing reports with R Markdown is also useful because if you find an error or want to modify the text or a figure later on, you can simply change the code and re-run the R Markdown file to get the updated output!
-
+(If you are familiar with Jupyter notebooks in the Python programming environment, R Markdown is  R's equivalent of a Jupyter notebook.)
 
 ## Creating a reports directory
 _[Back to top](#contents)_
 
 To get started, let's use the Unix Shell to create a directory within `un-report` called `reports` where we will write our reports to the UN.
-First, open the Unix Shell and `cd` to `un-report`
+First, open the Unix Shell and `cd` to `un-report`:
 
 ~~~
 pwd
@@ -79,6 +68,8 @@ mkdir reports
 /home/USERNAME/Desktop/un-report/reports/
 ~~~
 {: .output}
+
+Note that there is an option to use the terminal from R Studio (tab next to Console), but on Windows computers this terminal might not be a Unix Shell. 
 
 
 ## Creating an R Markdown file
[…]
