### 2021-11-22 cdce80c
Make sure everything is in a function
