### 2021-11-22 939090c
Actually run the detect-broken-links script 🤦🏻‍♀️
