### 2021-11-23 4aa16da
Consolidate R script steps
