### 2021-11-23 9d0c1dd
Revert "Consolidate R script steps"
This reverts commit 4aa16daa6f7aa3f630c6ebb2973a0178724683b3.
