+++ _episodes/01-r-plotting.md
@@ -31,6 +31,7 @@ keypoints:
 
 
 ### Contents
+
 1. [Introduction to R and RStudio](#introduction-to-r-and-rstudio)
 1. [Loading and reviewing data](#loading-and-reviewing-data)
 1. [Understanding commands](#understanding-commands)
@@ -44,8 +45,8 @@ keypoints:
     + [Plot themes](#plot-themes)
     + [Facets](#facets)
     + [Saving plots](#saving-plots)
-1. [Bonus topics!] 
-    + [Creating complex plots](#bonus-topics:-creating-complex-plots)
+1. [Bonus](#bonus)
+    + [Creating complex plots](#creating-complex-plots)
       + [Animated plots](#animated-plots)
       + [Map plots](#map-plots)
 1. [Glossary of terms](#glossary-of-terms)
@@ -55,7 +56,7 @@ keypoints:
 > > ## Solution:
 > > There are lots of different reasons, including to perform data analysis and generate figures. I'm sure you have morespecific reasons for why you'd like to learn!
 > {: .solution}
-{: challenge}
+{: .challenge}
 
 # Introduction to R and RStudio
 _[Back to top](#contents)_
@@ -84,7 +85,7 @@ To make your life in R easier, there is a great (and free!) program called RStud
 > > ## Solution:
 > > On some high-performance computer systems (e.g. Amazon Web Services) you typically can't get a display like RStudio to open. If you're at the University of Michigan and have access to Great Lakes, then you might want to learn more about [resources](https://arc-ts.umich.edu/greatlakes/user-guide/#document-3) to run RStudio on Great Lakes.
 > {: .solution}
-{: challenge}
+{: .challenge}
 
 > ## The tidyverse vs Base R
 > If you've used R before, you may have learned commands that are different than the ones we will be using during this workshop. We will be focusing on functions from the [tidyverse](https://www.tidyverse.org/). The "tidyverse" is a collection of R packages that have been designed to work well together and offer many convenient features that do not come with a fresh install of R (aka "base R"). These packages are very popular and have a lot of developer support including many staff members from RStudio. These functions generally help you to write code that is easier to read and maintain. We believe learning these tools will help you become more productive more quickly.
@@ -93,11 +94,11 @@ To make your life in R easier, there is a great (and free!) program called RStud
 
 To get started, we'll spend a little time getting familiar with the RStudio environment and setting it up to suit your tastes. When you start RStudio, you'll have three panels.
 
-![]({{ page.root }}/fig/r-plotting/initial_rstudio.png)
+<img src="{{ page.root }}/fig/r-plotting/initial_rstudio.png" width="600"/>
 
 On the left you'll have a panel with three tabs - Console, Terminal, and Jobs. The Console tab is what running R from the command line looks like. This is where you can enter R code. Try typing in `2+2` at the prompt (>). In the upper right panel are tabs indicating the Environment, History, and a few other things. If you click on the History tab, you'll see the command you ran at the R prompt.
 
-![]({{ page.root }}/fig/r-plotting/history.png)
+<img src="{{ page.root }}/fig/r-plotting/history.png" width="600"/>
 
 In the lower right panel are tabs for Files, Plots, Packages, Help, and Viewer. You used the Packages tab to install tidyverse.
 
@@ -105,13 +106,13 @@ We'll spend more time in each of these tabs as we go through the workshop, so we
 
 You might want to alter the appearance of your RStudio window. The default appearance has a white background with black text. If you go to the Tools menu at the top of your screen, you'll see a "Global options" menu at the bottom of the drop down; select that.
 
-![]({{ page.root }}/fig/r-plotting/global_options.png)
+<img src="{{ page.root }}/fig/r-plotting/global_options.png" width="600"/>
 
[…]
