+++ _episodes/01-r-plotting.md
@@ -94,11 +94,11 @@ To make your life in R easier, there is a great (and free!) program called RStud
 
 To get started, we'll spend a little time getting familiar with the RStudio environment and setting it up to suit your tastes. When you start RStudio, you'll have three panels.
 
-<img src="{{ page.root }}/fig/r-plotting/initial_rstudio.png" width="600"/>
+<img src="{{ page.root }}/fig/r-plotting/initial_rstudio.png" width="700"/>
 
 On the left you'll have a panel with three tabs - Console, Terminal, and Jobs. The Console tab is what running R from the command line looks like. This is where you can enter R code. Try typing in `2+2` at the prompt (>). In the upper right panel are tabs indicating the Environment, History, and a few other things. If you click on the History tab, you'll see the command you ran at the R prompt.
 
-<img src="{{ page.root }}/fig/r-plotting/history.png" width="600"/>
+<img src="{{ page.root }}/fig/r-plotting/history.png" width="700"/>
 
 In the lower right panel are tabs for Files, Plots, Packages, Help, and Viewer. You used the Packages tab to install tidyverse.
 
@@ -1641,9 +1641,13 @@ library(gifski)
 ~~~
 {: .language-r}
 
+
 > ## Reviewing how to create a scatter plot
 > **Part 1:**
-> Let's start by creating a static plot using `ggplot()`, as we've been doing so far. This time, lets put `log(gdpPercap)` on the x-axis, to help spread out our data points, and life expectancy on our y-axis. Also map the point size to the population of the country, and the color of the points to the continent.
+> Let's start by creating a static plot using `ggplot()`, as we've been doing so
+> far. This time, lets put `log(gdpPercap)` on the x-axis, to help spread out our
+> data points, and life expectancy on our y-axis. Also map the point size to the
+> population of the country, and the color of the points to the continent.
 >
 > > ## Solution
 > >
@@ -1713,65 +1717,70 @@ animatedHansPlot
 
 
 ~~~
-Rendering [--------------------------------------------] at 1.7 fps ~ eta: 1m
-Rendering [>-------------------------------------------] at 1.6 fps ~ eta: 1m
-Rendering [=>------------------------------------------] at 1.6 fps ~ eta: 1m
-Rendering [==>-----------------------------------------] at 1.6 fps ~ eta: 1m
-Rendering [===>----------------------------------------] at 1.6 fps ~ eta: 1m
-Rendering [====>---------------------------------------] at 1.6 fps ~ eta: 1m
-Rendering [=====>--------------------------------------] at 1.6 fps ~ eta: 1m
-Rendering [=====>--------------------------------------] at 1.5 fps ~ eta: 1m
-Rendering [======>-------------------------------------] at 1.4 fps ~ eta: 1m
+Rendering [--------------------------------------------] at 1.2 fps ~ eta: 1m
+Rendering [>------------------------------------------] at 0.92 fps ~ eta: 2m
+Rendering [>------------------------------------------] at 0.81 fps ~ eta: 2m
+Rendering [=>-----------------------------------------] at 0.91 fps ~ eta: 2m
+Rendering [=>-----------------------------------------] at 0.87 fps ~ eta: 2m
+Rendering [==>----------------------------------------] at 0.93 fps ~ eta: 2m
+Rendering [==>----------------------------------------] at 0.99 fps ~ eta: 2m
+Rendering [===>------------------------------------------] at 1 fps ~ eta: 1m
+Rendering [===>----------------------------------------] at 1.1 fps ~ eta: 1m
+Rendering [====>---------------------------------------] at 1.1 fps ~ eta: 1m
+Rendering [====>---------------------------------------] at 1.2 fps ~ eta: 1m
[…]
