+++ .github/workflows/website.yml
@@ -1,9 +1,6 @@
 name: Website
 on:
-  push:
-    branches:
-      - !gh-pages
-  pull_request: []
+  pull_request
 
 jobs:
   build:
