+++ README.md
@@ -1,5 +1,8 @@
-U-M SWC Curriculum
-==================
+# U-M SWC Curriculum
+
+[![Check](https://github.com/UMCarpentries/intro-curriculum-r/actions/workflows/check.yml/badge.svg)](https://github.com/UMCarpentries/intro-curriculum-r/actions/workflows/check.yml)
+[![License](https://img.shields.io/badge/License-MIT%2BCC--BY-blue)](LICENSE.md)
+[![DOI](http://img.shields.io/badge/DOI-10.1101/2021.06.15.448091-B31B1B.svg)](https://doi.org/10.1101/2021.06.15.448091)
 
 This is the custom curriculum developed by the University of Michigan Software Carpentry instance.
 It is in the `alpha` stage of development; we are hosting pilot workshops to gather feedback.
