+++ .github/workflows/website.yml
@@ -1,6 +1,5 @@
 name: Website
-on:
-  pull_request
+on: pull_request
 
 jobs:
   build:
