+++ .github/workflows/check-links.yml
@@ -33,7 +33,7 @@ jobs:
       - name: Check URL links
         run: |
           source("bin/detect-broken-links.R")
-          detect_broken_links("https://umcarpentries.org/intro-curriculum-r/00-intro/index.html",
+          detect_broken_links("https://umcarpentries.org/intro-curriculum-r/index.html",
             output_filename = "broken_links.tsv")
         shell: Rscript {0}
 
