### 2021-11-24 49a893d
Run the workflow every Monday at midnight
