### 2021-11-24 b76cbc0
Fix known broken links where possible
Currently the gganimate site is down
https://github.com/thomasp85/gganimate/issues/450
