+++ _episodes_rmd/01-r-plotting.Rmd
@@ -44,7 +44,7 @@ set.seed(20210109)
     + [Importing datasets](#importing-datasets)
     + [Discrete plots](#discrete-plots)
     + [Layers](#layers)
-    + [Color vs. Fill](#color-vs.-fill)
+    + [Color vs. Fill](#color-vs-fill)
     + [Univariate plots](#univariate-plots)
     + [Plot themes](#plot-themes)
     + [Facets](#facets)
@@ -68,7 +68,7 @@ _[Back to top](#contents)_
 In this session we will be testing the hypothesis that a country's life expectancy is related to the total value of its finished goods and services, also known as the Gross Domestic Product (GDP).
 To test this hypothesis, we'll need two things: data and a platform to analyze the data.
 
-You already [downloaded the data]({{ page.root }}/setup/). But what platform will we use to analyze the data? We have many options!
+You already [downloaded the data]({{ page.root }}/setup). But what platform will we use to analyze the data? We have many options!
 
 We could try to use a spreadsheet program like Microsoft Excel or Google sheets that have limited access, less flexibility, and don't easily allow for things that are critical to ["reproducible" research](https://journals.plos.org/ploscompbiol/article?id=10.1371/journal.pcbi.1003285), like easily sharing the steps used to explore and make changes to the original data.
 
@@ -80,9 +80,9 @@ Instead, we'll use a programming language to test our hypothesis. Today we will 
 >
 {: .solution}
 
-To run R, all you really need is the R program, which is available for computers running the Windows, Mac OS X, or Linux operating systems. You downloaded R while getting [set up]({{ page.root }}/setup/) for this workshop.
+To run R, all you really need is the R program, which is available for computers running the Windows, Mac OS X, or Linux operating systems. You downloaded R while getting [set up]({{ page.root }}/setup) for this workshop.
 
-To make your life in R easier, there is a great (and free!) program called RStudio that you also downloaded and used during [set up]({{ page.root }}/setup/). As we work today, we'll use features that are available in RStudio for writing and running code, managing projects, installing packages, getting help, and much more. It is important to remember that R and RStudio are different, but complementary programs. You need R to use RStudio.
+To make your life in R easier, there is a great (and free!) program called RStudio that you also downloaded and used during [set up]({{ page.root }}/setup). As we work today, we'll use features that are available in RStudio for writing and running code, managing projects, installing packages, getting help, and much more. It is important to remember that R and RStudio are different, but complementary programs. You need R to use RStudio.
 
 > Bonus Exercise: Can you think of a reason you might not want to use RStudio?
 >
@@ -182,7 +182,7 @@ library(tidyverse)
 > message tells us what packages were installed when we loaded `tidyverse` (these
 > include `ggplot2`, which we'll be using a lot in this lesson, and `dyplr`, which
 > you'll be introduced to tomorrow in the 
-> [R for Data Analysis lesson](05-r-markdown.Rmd)).
+> [R for Data Analysis lesson]({{ page.root }}/05-r-markdown)).
 >
 > The second section of messages gives a list of "conflicts." Sometimes, the
 > same function name will be used in two different packages, and R has to decide
@@ -961,7 +961,7 @@ Both `geom_violin` and `geom_jitter` will inherit the default values of `aes(con
 {: .solution}
 
 
-## Color vs. Fill
+## Color vs. Fill 
 _[Back to top](#contents)_
 
 Let's say we want to spice up our plot a bit by adding some color. Maybe we want our violin color to a fancy color like "darkolivegreen." We can do this by explicitly setting the color aesthetic inside the `geom_violin` function. Note that because we are assigning a color directly and not using any values from our data to do so, we do not need to use the `aes()` mapping function. Let's try it out:
