+++ _episodes/06-conclusion.md
@@ -96,7 +96,7 @@ In the following, we list some strategies and resources we find useful. As you m
 
 * [RStudio Community Forum](https://community.rstudio.com/). This is a site that is welcoming to new R users where you can ask R programming questions.
 * [Search R Package Documentation](https://rdrr.io/). This site is very useful when you see someone using a function but don't know which package it comes from. This site lets you search across the help pages for all R packages.
-* [RStudio Cheat Sheets](https://rstudio.com/resources/cheatsheets/). A good collection of cheat sheets to print out and hang at your desk. For example: [base R](https://rstudio.com/wp-content/uploads/2016/10/r-cheat-sheet-3.pdf), [tidyverse](https://s3.amazonaws.com/assets.datacamp.com/blog_assets/Tidyverse+Cheat+Sheet.pdf), and [ggplot](https://github.com/rstudio/cheatsheets/raw/master/data-visualization-2.1.pdf)
+* [RStudio Cheat Sheets](https://rstudio.com/resources/cheatsheets/). A good collection of cheat sheets to print out and hang at your desk. For example: [base R](https://github.com/rstudio/cheatsheets/blob/main/base-r.pdf), [tidyverse](https://s3.amazonaws.com/assets.datacamp.com/blog_assets/Tidyverse+Cheat+Sheet.pdf), and [ggplot](https://raw.githubusercontent.com/rstudio/cheatsheets/main/data-visualization.pdf)
 * If there is a specific R package that you want to use (e.g. [mikropml](http://www.schlosslab.org/mikropml/) for machine learning analysis), you can use the vignettes to help you learn the basics of the package. You can search for the vignette online, or use the `vignette()` function in R to open them.
 
 ### R Markdown
