+++ README.md
@@ -4,8 +4,10 @@
 [![License](https://img.shields.io/badge/License-MIT%2BCC--BY-blue)](LICENSE.md)
 [![DOI](http://img.shields.io/badge/DOI-10.1101/2021.06.15.448091-B31B1B.svg)](https://doi.org/10.1101/2021.06.15.448091)
 
-This is the custom curriculum developed by the University of Michigan Software Carpentry instance.
-It is in the `alpha` stage of development; we are hosting pilot workshops to gather feedback.
+This is the custom curriculum developed by the University of Michigan Software
+Carpentry instance.
+It is in the `beta` stage of development; we are continuing to host workshops to
+gather feedback and polish the curriculum.
 
 The rendered website can be found [here](https://umcarpentries.github.io/intro-curriculum-r).
 
