### 2021-12-30 bff3841
Suggesting changes to CoC guidelines language
Suggesting changes to CoC reporting language. Since this isn't in a Carpentries repository, the reporting guidelines would just apply when it's a registered Carpentries workshop. Otherwise, The Carpentries wouldn't necessarily have any knowledge or control over what happened in this space. If you have some other reporting structure to include in other situations, you could put that here too.
