+++ CODE_OF_CONDUCT.md
@@ -2,10 +2,10 @@
 title: "Contributor Code of Conduct"
 ---
 As contributors and maintainers of this project,
-we pledge to follow the [Carpentry Code of Conduct][coc].
+we pledge to follow the [The Carpentries Code of Conduct][coc].
 
-Instances of abusive, harassing, or otherwise unacceptable behavior
-may be reported by following our [reporting guidelines][coc-reporting].
+If this lesson is being taught as a registered Carpentries workshop, instances of abusive, harassing, or otherwise unacceptable behavior
+may be reported by following their [reporting guidelines][coc-reporting].
 
 [coc]: https://docs.carpentries.org/topic_folders/policies/code-of-conduct.html
 [coc-reporting]: https://docs.carpentries.org/topic_folders/policies/incident-reporting.html
