### 2021-12-30 UMCarpentries/intro-curriculum-r#170 @tracykteal
Clarification on connection to Carpentries lessons
This is a lesson based on Software Carpentry and developed by the UM Carpentries instance, so it makes sense to refer to it as 'Carpentries'. It's not officially Carpentries though, given that it's not in a Carpentries space, so I think some clarification on that point is needed to avoid confusion.

One option would be to amend the intro on the index page to be something like:

This is a custom curriculum developed by the University of Michigan Carpentries instance. While it is based on The Carpentries curriculum and teaching practices, it is not officially overseen by The Carpentries organization. We welcome any feedback or questions to us through [GitHub](https://github.com/UMCarpentries/intro-curriculum-r) or [other contact information].

You maybe already had some conversations with Carpentries about naming and acknowledgement though, so if so, please just use whatever they've recommended!
