### 2021-12-30 #144 @tracykteal
Thanks! I've gone through my review, and I'm very sorry for the delay. I also recommend this paper for publication without revision. This is an excellent lesson that creates a clear 2 day curriculum that is easy to use for instructors and learners, reducing the overhead of getting started.

This is a lesson based on Software Carpentry and developed by the UM Carpentries instance, so it makes sense to refer to it as 'Carpentries'. It's not officially Carpentries though, given that it's not in a Carpentries space, so some of the references to Carpentries infrastructure aren't relevant. I put in one PR to address.

- [ ] https://github.com/UMCarpentries/intro-curriculum-r/pull/169

I also created an issue with some potential clarification language on the index page.

- [ ] https://github.com/UMCarpentries/intro-curriculum-r/issues/170

The authors may have already had some conversations with Carpentries about naming and acknowledgement though, so if so, please just use whatever they've recommended!
