### 2021-12-31 #144 @zenalapp
Thanks, @tracykteal! We have merged your PR and updated the language about our connection to Carpentries lessons.

@magsol is there anything else we need to do?
