### 2022-01-01 #144 @magsol
@zenalapp This looks great! At this point I'd ask for you to create a *tagged release* (only if you don't already have one, or have made updates since your last one) and *archive*, and post both the release **version number** and the **archive DOI** here, and I can take it from there.
