### 2022-01-01 #144 @kelly-sovacool
@magsol I just released version `1.0.0` with DOI `10.5281/zenodo.5812479`.
