### 2022-01-01 #144 @labarba
@pschloss – Could you try to fix the horizontal badness on URLs that appear in the References? Maybe using a `\url` inside the `howpublished` field of the `.bib` entry?
