### 2022-01-02 f0c523d
Use \url{} in howpublished field of bibtex entries
https://github.com/openjournals/jose-reviews/issues/144#issuecomment-1003627447
