### 2022-01-03 #144 @zenalapp
Just merged the PR from @kelly-sovacool to fix this.
