### 2022-01-08 #144 @labarba
Congratulations, @zenalapp and co-authors, your article is published in JOSE! 🚀

Huge thanks to our Editor: @magsol and the Reviewers: @kdillmcfarland, @tracykteal — we appreciate your contribution to this adventure in publishing 🙏
